# api/v2/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/profile": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Позволяет пользователю получить информацию о своем аккаунте со всеми правами и ролями",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Профиль",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Позволяет обновить данные по своему профилю",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Изменение профиля",
                "parameters": [
                    {
                        "description": "Обновленные данные профиля",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ProfileUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profile/addresses": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Позволяет получить информацию об адресах всех получателей привязанных к пользователю",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gitea.technodom.kz_intechno_techno-sso_internal_ports_http_resources_profile_v1.AddressesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/profile/password": {
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Позволяет обновить пользователю пароль",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Смена пароля",
                "parameters": [
                    {
                        "description": "Новый пароль",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdatePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/profile/receivers": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Позволяет получить информацию(кроме адреса) обо всех получателях привязанных к пользователю",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gitea.technodom.kz_intechno_techno-sso_internal_ports_http_resources_profile_v1.ReceiversRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ProfileResponse": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "iin": {
                    "type": "integer"
                },
                "lang": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "phone": {
                    "type": "string"
                },
                "phones": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "receivers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Receiver"
                    }
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sex": {
                    "type": "string"
                },
                "tdid": {
                    "type": "string"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
...
        "api.ProfileUpdateRequest": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "iin": {
                    "type": "integer"
                },
                "lang": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "sex": {
                    "type": "string"
                }
            }
        },
...
        "api.UpdatePasswordRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
...
        "gitea.technodom.kz_intechno_techno-sso_internal_ports_http_resources_profile_v1.AddressesResponse": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ReceiverAddressResponse"
                    }
                }
            }
        },
        "gitea.technodom.kz_intechno_techno-sso_internal_ports_http_resources_profile_v1.ReceiversRequest": {
            "type": "object",
            "properties": {
                "receivers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ReceiverResponse"
                    }
                }
            }
        },
...
        "models.AddressGeo": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "string"
                },
                "lng": {
                    "type": "string"
                }
            }
        },
...
        "models.ErrorDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "application-определенный код ошибки",
                    "type": "integer"
                },
                "message": {
                    "description": "application-level сообщение, для дебага",
                    "type": "string"
                },
                "status": {
                    "description": "сообщение пользовательского уровня",
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/models.ErrorDetails"
                },
                "validation": {
                    "description": "ошибки валидации",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "models.Organization": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "bik": {
                    "type": "string"
                },
                "bin": {
                    "type": "string"
                },
                "iic": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Receiver": {
            "type": "object",
            "properties": {
                "additionalPhone": {
                    "type": "string"
                },
                "address": {
                    "$ref": "#/definitions/models.ReceiverAddress"
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isDefault": {
                    "type": "boolean"
                },
                "isOrganization": {
                    "type": "boolean"
                },
                "lastname": {
                    "type": "string"
                },
                "organization": {
                    "$ref": "#/definitions/models.Organization"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "models.ReceiverAddress": {
            "type": "object",
            "properties": {
                "apartment": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "floor": {
                    "type": "string"
                },
                "geo": {
                    "$ref": "#/definitions/models.AddressGeo"
                },
                "house": {
                    "type": "string"
                },
                "region": {
                    "$ref": "#/definitions/models.ReceiverRegion"
                },
                "street": {
                    "type": "string"
                },
                "zipcode": {
                    "type": "string"
                }
            }
        },
        "models.ReceiverAddressResponse": {
            "type": "object",
            "properties": {
                "apartment": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "floor": {
                    "type": "string"
                },
                "geo": {
                    "$ref": "#/definitions/models.AddressGeo"
                },
                "house": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "region": {
                    "$ref": "#/definitions/models.ReceiverRegion"
                },
                "street": {
                    "type": "string"
                },
                "zipcode": {
                    "type": "string"
                }
            }
        },
        "models.ReceiverRegion": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.ReceiverResponse": {
            "type": "object",
            "properties": {
                "additionalPhone": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isDefault": {
                    "type": "boolean"
                },
                "isOrganization": {
                    "type": "boolean"
                },
                "lastname": {
                    "type": "string"
                },
                "organization": {
                    "$ref": "#/definitions/models.Organization"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
    }
}